Model definitions have increased substantially.  Testing almost complete of DB changes.

+++ api/models/concession.js
@@ -8,7 +8,7 @@ module.exports = function(sequelize, DataTypes) {
 			defaultValue: ''
 		},
 		price: {
-			type: DataTypes.FLOAT,
+			type: DataTypes.DECIMAL(6,2),
 			allowNull: true
 		},
 		description: {
